Dto improvements (#3)
* TidyUp: 1. Make sure we're not exposing user password in UserDto

* TidyUp: 5. Introduce (and apply) spotless plugin

* TidyUp: 2. BikeController: "return ResponseEntity.ok(BikeDto.fromEntity(bikeService.addNewBike(bikeDto.toEntity())));" looks a bit shit.

* TidyUp: 3. Currently, in controllers, we wrap everything in ResponseEntity. Unwrap unnecessary wrapping

* 4. Create GHAs for testing etc

* 4. Create GHAs for testing etc

+++ src/main/java/com/bike/service/BikeService.java
@@ -1,61 +1,62 @@
 package com.bike.service;
 
+import static java.time.LocalDate.now;
+
 import com.bike.model.Bike;
 import com.bike.model.BikeHire;
 import com.bike.repository.BikeHireRepository;
 import com.bike.repository.BikeRepository;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
-
-import static java.time.LocalDate.now;
-import static java.util.stream.Collectors.toList;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 @Slf4j
 @Service
 @Transactional
 public class BikeService {
 
-    private final BikeRepository bikeRepository;
-    private final BikeHireRepository bikeHireRepository;
+  private final BikeRepository bikeRepository;
+  private final BikeHireRepository bikeHireRepository;
 
-    @Autowired
-    public BikeService(BikeRepository bikeRepository,
-                       BikeHireRepository bikeHireRepository) {
-        this.bikeRepository = bikeRepository;
-        this.bikeHireRepository = bikeHireRepository;
-    }
+  @Autowired
+  public BikeService(BikeRepository bikeRepository, BikeHireRepository bikeHireRepository) {
+    this.bikeRepository = bikeRepository;
+    this.bikeHireRepository = bikeHireRepository;
+  }
 
-    public List<Bike> getList() {
-        return bikeRepository.findAll();
-    }
+  public List<Bike> getList() {
+    return bikeRepository.findAll();
+  }
 
-    public Bike getById(UUID id) throws EntityNotFoundException {
-        return bikeRepository.findById(id).orElseThrow(() ->
-                new EntityNotFoundException("Unable to find com.bike.model.Bike with id " + id));
-    }
+  public Bike getById(UUID id) throws EntityNotFoundException {
+    return bikeRepository
+        .findById(id)
+        .orElseThrow(
+            () -> new EntityNotFoundException("Unable to find com.bike.model.Bike with id " + id));
+  }
 
-    public Bike addNewBike(Bike bike) {
-        return bikeRepository.save(bike);
-    }
+  public Bike addNewBike(Bike bike) {
+    return bikeRepository.save(bike);
+  }
 
-    public void deleteBike(UUID bikeId) {
-        bikeRepository.deleteById(bikeId);
+  public void deleteBike(UUID bikeId) {
+    if (!bikeRepository.existsById(bikeId)) {
+      throw new EntityNotFoundException("Unable to find Bike with id " + bikeId);
     }
+    bikeRepository.deleteById(bikeId);
+  }
 
-    public List<BikeHire> getHiresForBike(Bike bike) {
-        return bikeHireRepository.findAllByBike(bike);
-    }
+  public List<BikeHire> getHiresForBike(Bike bike) {
+    return bikeHireRepository.findAllByBike(bike);
+  }
 
-    public List<BikeHire> getFutureHiresForBike(Bike bike) {
-        LocalDate today = now();
-        return bikeHireRepository.findAllByBikeAndStartDateNowOrLater(bike, today);
-    }
+  public List<BikeHire> getFutureHiresForBike(Bike bike) {
+    LocalDate today = now();
+    return bikeHireRepository.findAllByBikeAndStartDateNowOrLater(bike, today);
+  }
 }
